Nudge transitions if not animating
With the recent changes on Map::renderSync(), we now must call for
Map::nudgeTransitions() in order to check if there are any pending
transitions and/or animations to be completed.

However, if we do this while animating via client-side (i.e. using
QPropertyAnimation), there is a perceived performance decrease due to
the fact that the view can end up being updated twice for each new
lat/lon/zoom value. We check if there is any animation happening at the
current time point before calling Map::nudgeTransitions().